+++ armada_backend/api_list.py
@@ -5,7 +5,8 @@ from distutils.util import strtobool
 import web
 
 from armada_backend import api_base
-from armada_backend.utils import get_ship_name, get_logger
+from armada_backend.models.ships import get_ship_name
+from armada_backend.utils import get_logger
 from armada_command.consul import kv
 from armada_command.consul.consul import consul_query
 
@@ -36,7 +37,7 @@ def __create_dict_from_tags(tags):
 
 
 def _get_services_list(filter_microservice_name, filter_env, filter_app_id, filter_local):
-    consul_key = 'containers_parameters_list'
+    consul_key = 'services'
     if filter_local:
         consul_key = '{}/{}'.format(consul_key, get_ship_name())
 
@@ -45,11 +46,7 @@ def _get_services_list(filter_microservice_name, filter_env, filter_app_id, filt
     if not services_by_ship:
         return {}
 
-    result = {}
-    for services_dict in services_by_ship.values():
-        result.update(_parse_single_ship(services_dict, filter_microservice_name, filter_env, filter_app_id))
-
-    return result
+    return _parse_single_ship(services_by_ship, filter_microservice_name, filter_env, filter_app_id)
 
 
 def _get_running_services(filter_microservice_name, filter_env, filter_app_id, filter_local):
@@ -121,7 +118,7 @@ def _parse_single_ship(services_dict, filter_microservice_name, filter_env, filt
         return result
 
     if filter_microservice_name:
-        services_list = fnmatch.filter(services_list, 'ships/*/service/{}/*'.format(filter_microservice_name))
+        services_list = fnmatch.filter(services_list, 'services/*/{}/*'.format(filter_microservice_name))
 
     for service in services_list:
         service_dict = services_dict[service]